Remove the C++ guard as all of the C standard headers behave well in C++
