+++ src/inc/dnmd.h
@@ -3,9 +3,6 @@
 
 #include <stdint.h>
 #include <stddef.h>
-
-#ifndef __cplusplus
-
 #include <stdbool.h>
 // MacOS doesn't have uchar.h
 #if defined(__has_include)
@@ -14,7 +11,6 @@
 #else
 // When uchar.h isn't available, define char16_t as per the C standard.
 typedef uint_least16_t char16_t;
-#endif
 #endif
 
 #endif
