+++ src/inc/internal/dnmd_platform.hpp
@@ -13,10 +13,26 @@
 
 #endif // !BUILD_WINDOWS
 
+#include <cstdlib>
+
 #include <dncp.h>
 #include <cor.h>
 #include <dnmd.hpp>
 
 #define ARRAY_SIZE(a) (sizeof(a) / sizeof(*a))
+
+template<typename T>
+struct malloc_deleter_t
+{
+    using pointer = T*;
+    void operator()(T* mem)
+    {
+        ::free((void*)mem);
+    }
+};
+
+// C++ lifetime wrapper for malloc'd memory
+template<typename T>
+using malloc_ptr = std::unique_ptr<T, malloc_deleter_t<T>>;
 
 #endif // _SRC_INC_INTERNAL_DNMD_PLATFORM_HPP_
